requirements.txt
faiss-cpu
groq
pdfplumber
sentence-transformers
requests